Add single gravity and split flip
Allows hybridGravity to fulfill all the things needed for in Discussion 3, running all the stars in one gravity code, the other, flipping the mass cut, and doing the split.

Still need to get the half mass and core-radii somehow. All other information  is built in as needed for now. Also need to set the virial radius somehow.